backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
bcrypt
pyjwt
pytest
httpx
requests
gunicorn==21.2.0
